requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cryptography==44.0.3
fastapi==0.115.12
h11==0.16.0
idna==3.10
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
numpy==2.2.5
opencv-python-headless==4.11.0.86
packaging==25.0
pdf2image==1.17.0
pdfminer.six==20250506
pillow==11.2.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pytesseract==0.3.13
python-docx==1.1.2
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
